refactor: split three files back under the per-file line limit
This branch pushed three files past the repo's 800-line limit, so CI's
pre-commit job would have failed. Split by concern rather than by cutting
lines to fit:

* `ad_completeness_review` (1123) -> 787, with two siblings:
  - `ad_completeness_progress` — convergence accounting. Everything there
    answers "is the agent still making progress?" and nothing reads a
    report. It also now owns the stall bookkeeping the reviewer used to do
    inline, and a public `clear_all()`.
  - `ad_completeness_rules` — the report's SHAPE (the regexes) plus the
    two checks that are decidable from the document alone: per-campaign
    spend against the platform export, and the cross-cutting gaps that
    belong to the run rather than to any one marketplace.
  The reviewer re-exports the public names, so the stop hook and task
  runner import exactly what they did before.

* `routers/tasks.py` (866) -> 781. The submission-resolution block moved
  to `task_submission`, which was itself split out of tasks.py for this
  reason and is the right home: pointer / dangling-pointer / narrated
  deliverable are one decision about what text the gates should grade.
  The generic soft-gate loop moved with it. `resolve_submitted_result`
  returns `(text, came_from_a_file)` because the endpoint reports the
  second half as `resolved_from_file` and a contract test pins that key —
  returning only the text would have broken the API silently.

* `amazon-ads/scripts/ads_bulk.py` (818) -> 685, with the 52-column sheet
  schema in `ads_schema.py` beside it (the pattern `amazon-listing`
  already uses). Added to MANIFEST.txt so it syncs to the agent's
  workspace; the script's own directory is on sys.path when it runs, so
  the sibling import resolves the same way in the agent's environment.

Two test-side consequences, both fixed as design rather than patched:

* The skill-script loader now puts the script's directory on sys.path,
  exactly as invoking it does — otherwise the sibling import would fail
  only under test.
* The stop-gate fixtures used to clear seven module globals by name, so
  any counter added later would have leaked between tests until a
  spurious pass gave it away. They call `clear_all()` instead.

No behaviour change intended: unit 1633 pass, workflow pass, and the full
`pre-commit run --all-files` is green including the line-limit hook.
