+++ app/skills_v2/MANIFEST.txt
@@ -15,6 +15,7 @@ amazon-ads/references/output-spec.md
 amazon-ads/references/reviewer-loop.md
 amazon-ads/references/tuning-history.md
 amazon-ads/scripts/ads_bulk.py
+amazon-ads/scripts/ads_schema.py
 amazon-ads/scripts/md_to_pdf.py
 amazon-ads/requirements.txt
 amazon-listing/SKILL.md
